Moved Sub Tier methods to a separate class
