+++ src/main/java/com/gospelbg/Utils.java
@@ -1,0 +1,35 @@
+package com.gospelbg;
+
+import com.github.twitch4j.common.enums.SubscriptionPlan;
+
+public class Utils {
+     public static String PlanToString(SubscriptionPlan plan) {
+        switch (plan.toString()) {
+            case "Prime":
+                return "Tier 1 (Prime)";
+            case "1000":
+                return "Tier 1";
+            case "2000":
+                return "Tier 2";
+            case "3000":
+                return "Tier 3";
+            default:
+                return null;
+        }
+    }
+
+    public static String PlanToConfig(SubscriptionPlan plan) {
+        switch (plan.toString()) {
+            case "Prime":
+                return "TWITCH_PRIME";
+            case "1000":
+                return "TIER1";
+            case "2000":
+                return "TIER2";
+            case "3000":
+                return "TIER3";
+            default:
+                return null;
+        }
+    }
+}
